refactor(eve): extract the engine-neutral turn-step operation
turnStep's body moves to execution/turn-step-operation.ts with every
engine-owned capability injected: the pre-read durable session, the
callback base URL (previously resolved via getWorkflowMetadata inside the
body), the runtime constructor for delegated child runs, and a new
writeEveAttributes sink. The workflow-steps turnStep becomes a thin
"use step" shell that supplies those from the Workflow engine.

The attribute seam also removes the harness's last engine import:
tool-loop no longer calls setEveAttributes directly; the writer arrives
through ToolLoopHarnessConfig and node-step, and engines without an
attribute store simply omit it.

Zero behavior change: the body is verbatim, and the full unit and
integration tiers pass unmodified. First step of the loop-decoupling
migration (research/loop-interface.md): the operation is the seam the
loop programs' generate effect will call.
